+++ config/routes.rb
@@ -1,5 +1,11 @@
 Rails.application.routes.draw do
 
+  namespace :admin do
+    resources :team_members
+    resources :partners
+    resources :clinical_trials
+  end
+
   namespace :builder do
 
     resources :surveys do